[Cashtab][Move Msg Switch pt3] Add transition effect to address inputs
Summary: Related to T2772. Depends on D12525. Added transition effect to DestinationAddressSingle and DestinationAddressMulti. When user clicks the Multiple Recipients switch, the max height and opacity transitions on both the Single and Multi inputs, and the position of the DestinationMulti input is also transitioned.

Test Plan:
cd web/cashtab && npm start
navigate to send screen
toggle the multiple recipients switch
observe that the incoming input does not 'pop' on to the screen, but rather transitions

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12536