[Cashtab] Parse edge case eToken tx
Summary:
T2795

A new eToken swap wallet creates a type of tx previously ignored in Cashtab tx parsing. This diff parses it as an incoming XEC tx.

A task has been created to improve this parsing by flagging it as an eToken tx with an XEC output. In the future, we may be able to tag these txs as being from a specific app.

Test Plan:
- Ping me on telegram for the seed to a wallet with this kind of tx
- `npm start` and verify correct parsing

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D12549