[Cashtab] Fallback to block time if timeFirstSeen is zero
Summary:
T2806

If chronik has not seen a tx, timeFirstSeen will be zero and Cashtab will parse the tx as having been sent in 1970.

Fall back to blocktime in this case to get a mostly accurate time estimate.

Test Plan:
Ping me on Telegram for a wallet seed with some txs in this category
Import at cashtab.com and note wrong dates
`npm start` and confirm accurate dates

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12648