[Cashtab] [Deprecate toLegacyCash p1] Stop converting ecash addresses to bitcoincash in Send.js
Summary:
T2730

Part 1 of a stacked diff in support of deprecating the BCH param and address conversion in Cashtab

In this diff, we stop converting `ecash:` format addresses to `bitcoincash:` format. This was an artifact from when bch-js was used. With our own Transaction Builder and chronik, this is no longer necessary.

Test Plan:
`npm start`
Send an XEC tx

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12655