[Cashtab] [Simplify wallet structure p5] Remove function specific variable name change
Summary:
T2663

Depends on D12625

There is a case to be made for renaming `slpUtxos` and `nonSlpUtxos` to `tokenUtxos` and `xecUtxos` everywhere, but in this case I think it's better to stick to the original syntax. At any rate, we should be consistent about the name everywhere in the app.

Test Plan:
Confirm diff is changing variable names
Confirm I didn't mix them up and that tokenUtxos became slpUtxos, and xecUtxos became nonSlpUtxos
`npm test`
`npm start`, send a token tx, burn a token

Reviewers: #bitcoin_abc, kieran709

Reviewed By: #bitcoin_abc, kieran709

Differential Revision: https://reviews.bitcoinabc.org/D12659