[Cashtab] Deprecate BCH param from generateTxOutput function
Summary:
T2730

Depends on D12664

Deprecate BCH param from generateTxOutput function. The param is there only as an artifact and is not used in the function.

Test Plan:
`npm test`
`grep -r generatTxOutput src/` and make sure the param is removed from all instances of the function
`npm start` and send an XEC tx

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D12665