[Cashtab] [deprecate BCH from token send p3] Remove from createToken and related components
Summary:
T2730

Depends on D12690

Remove the BCH param from createToken. It is no longer used by the function and is only an artifact.

Also remove in all calls of createToken. The BCH param needs to be removed from this screen in this diff as it is no longer used by any other function there.

Test Plan:
`npm start`
Mint an etoken
`grep -r createToken src/` and confirm it is not called anywhere with the BCH param

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12691