[Linter] Fix MyPy linter with version 0.991
Summary:
Since this version, the linter returns a note each time it encounters a function with no annotation. This leads to an incredible amount of messages for no reason, and since we don't bother filtering by severity in the PHP parser, the same amount of lint errors.
This diff just ignores this message and everything is green again.

Test Plan:
With mypy 0.991:
  arc lint --everything
Check there is not a gazillion errors

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12726