Move CBlockFileInfo::ToString method where class is declared
Summary:
> CBlockFileInfo class is declared in src/chain.h, so move ToString
> definition to src/chain.cpp instead of src/node/blockstorage.cpp

`CBlockFileInfo` was moved to its own header file in D1654, so "where class is declared" is a new blockfileinfo.cpp file.
The class is used in txdb.h and validation.h, which are both in the `server` library, so put this new file in this library as well.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/23497 | core#23497]]
https://github.com/bitcoin/bitcoin/pull/23497/commits/e5b6aef61221b621ad77b5f075a16897e08835bf

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12768