Fix automated timings updates
Summary:
Chronik needs to be enabled so the chronik_* tests are accounted for.

The USDT tests are not designed to be run as part of a normal test framework run, as they need some extra privileges. As such it makes no sense to store their timing: as soon as the appriated conditions are met, they need to run no matter the duration.
Also exclude them from the automated timing.json update script check for the same reason.

Test Plan:
  mkdir buildTracing
  cmake -GNinja .. -DENABLE_TRACING
  ninja
  sudo ./test/functional/test_runner.py interface_usdt_net
Check the timing is not written to buildTracing/timings.json

Check it succeeds:
  BUILD_DIR=$PWD/../buildTImings ../contrib/source-control-tools/patch-recipes/update-timings.sh

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12826