[eCash dev libraries] [p1 ecash-ops] Starting from existing module
Summary:
T2847

Cloning in the git repo of the dependency as it is currently used by `ecashjs-lib`

Test Plan:
This is a starting point. Later diffs in the stack will follow plan outlined by task.

To test, confirm that this diff is adding in the repo found here unchanged, https://github.com/Bitcoin-com/bitcoincash-ops

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D12855