[Cashtab Docs] Removing docs from the monorepo
Summary:
T2841

Remove the docs subdirectory from the monorepo. docs.cashtab.com is currently unavailable and 301s to cashtab. Removing this from the monorepo as it is not being maintained.

Test Plan: Confirm this diff removes the `docs/` directory. Confirm in slack that this diff and D12918 can be landed i.e. confirm teamcity builds are configured correctly to take this change.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12919