Use CTransactionRef where applicable in coins_tests
Summary:
This avoids doing copies of CTransaction in the tests. This will make the test to work with a non copyable CTransaction.

Also removed the forward declaration of the UpdateCoins function in favor of a proper include. This will prevent the test from failing to build when the prototype changes.

Test Plan:
  ninja check-bitcoin-coins_tests

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12920