[Cashtab] Improve vertical centering of QRCode in Receive.js
Summary: Related to T2814. QRCode has been given a margin-top of 15%. If you think it would look better, I would stack this diff with one which moves the new InfoCtn above the QRCode so that all content is visible above the fold.

Test Plan:
cd web/cashtab && npm start
navigate to Receive screen
observe that the QRCode component is pushed further towards the center of the screen

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D12922