[cashtab] Ignore eslint false positive for jest import
Summary:
From eslint:
```
Jest is automatically in scope. Do not import "jest", as Jest doesn't export anything.
```

However this only stands for test files, jest is not automatically in the scope of other files and this causes a false positive, which can be safely ignored with an eslint single line error ignre comment.

Test Plan:
  cd web/cashtab
  npm ci
  npm test
  eslint scripts/test.js

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D12924