Bump the p2p_blockfilters initial block sync timeout
Summary: The generate call now implies a chain sync, but despite the RPC has a long timeout, the sync timeout is the default 60s which is not enough on some of our slow machines, especially the qemu ARM machine. This diff bump this timeout to prevent spurious CI failures.

Test Plan:
  ./test/functional/test_runner.py p2p_blockfilters

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12940