[avalanche] Add the local proof verification status details to getavalancheinfo
Summary:
The getavalancheinfo RPC returns whether the local proof is verified or not, but gives no indication about why it's not verified. Since the node waits for a block to trigger the verification this can lead to confusion when a user restarts and see his proof being unverified.

This diff adds another status field when the verified field is false to indicate whether the verification is pending or if the verification ocurred but failed.

Test Plan:
  ./test/functional/test_runner.py abc_rpc_getavalancheinfo

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12941