refactor: Add BlockManager getters
Summary:
A dependency of a dependency of an assumeutxo related backport uses `IsPruneMode`, so let's backport these getters.
There are some callsites in init.cpp missing in this diff due to a missing backport related to a warning when the disk space is low.There is no risk of breaking anything if this is backported later and it does not seem very important.

> Add BlockManager::LoadingBlocks()

https://github.com/bitcoin/bitcoin/pull/26900/commits/fa0f0436d83288262d7d764b1d239c1a6de6146f

> Add BlockManager::GetPruneTarget()

https://github.com/bitcoin/bitcoin/pull/26900/commits/fae71fe27ec021583aaeac09aa924522bb63db05

> Add BlockManager::IsPruneMode()

https://github.com/bitcoin/bitcoin/pull/26900/commits/faf7b4f1fc35f1488567e0e4a57ecb348596b992

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/26900 | core#26900]]

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15003