Update copyright year for explorer
Summary: Updating the footer copyright year to 2023. There was mention of using the CMakeLists.txt file in the monorepo as a better solution, but I dont know how to implement that. Can do if someone has some ideas

Test Plan: Run the app and see if the year is right

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12974