Remove reindex special case from the progress bar label
Summary:
This is a minor change in the status bar message when starting bitcoin-qt with the reindex flag.

The user knows which option they passed to the program, so it seems overly verbose to offer the user feedback whether or not they passed -reindex. Treat it as DISK, like all other cases that are treated as DISK:
 -reindex-chainstate
 -loadblock

This is a backport of [[https://github.com/bitcoin-core/gui/pull/697 | core-gui#697]]

Depends on D16004

Test Plan:
`ninja all check-all`

`src/qt/bitcoin-qt -reindex -regtest`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16006