[avalanche] Improve test coverage of availability scoring
Summary:
The good: The current unit test checks defaults used in production.
The bad: The for loops in this test run only once since defaults selected are step = tau.
The ugly: We don't know if this test actually works when step < tau.

We expect to be able to tweak the Avalanche statistics constants without unexpectedly breaking tests,
so lets make sure we can do that by testing that these loops do what we think they do when step < tau.

Depends on D13003

Test Plan:
  ninja check check-functional

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13004