[avalanche] Using C++17 structured bindings to make unit tests more readable
Summary:
See title. In the past we have been making a tradeoff where we increase test
coverage at the expense of test readability. No more of that since we can have
it all using C++17 structured bindings.

Depends on D13001 only because it touches some of the same code.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13002