wallet: drop anti-fee-sniping feature, set locktime = 0
Summary:
Post-consensus mitigates fee-sniping, so we don't need to set a locktime to the current block height when creating a wallet transaction.

This reduces the wallet fingerprintability, as light wallets don't have access to the current block height and cannot set the locktime.

The plan is to disable this feature by default also in Electrum ABC (but keep the option to do it for the sake of splitting coins).

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13059