[consensus] Move miner fund logic into its own function
Summary:
This will help simplify the migration to miner fund as policy. There is no
change in behavior.

The ideal design would not pull in fees as a function parameter but instead be
able to determine the fees from the block itself. This is inefficient to do
right now so we take advantage of the fact that the caller has already done
this calculation.

Test Plan:
  ninja check check-functional

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13071