deprecate softforks from getblockchaininfo RPC
Summary:
This removes a bunch of code related to the currently empty "softforks" output of the getblockchain RPCs.
The field itself is kept (DEPRECATED) but the value will now always be empty. This only makes a difference on testnet for the testdummy deployment (which will be removed in D13158).

Depends on D13156

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13157