[tests] show the proofid when representing a proof
Summary: This would have made it easier to debug the issue fixed by D13099, to verify that the received `msg_avaproof` was indeed the one corresponding to the sent `msg_getdata` before comparing the timestamps.

Test Plan: Make a test involving a `msg_avaproof` (e.g. abc_p2p_compactproofs) fail  and inspect the log for a `msg_avaproof`.

Reviewers: #bitcoin_abc, sdulfari, Fabien

Reviewed By: #bitcoin_abc, sdulfari, Fabien

Subscribers: Fabien, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D13100