[explorer] Fix table sorting issues
Summary:
Sorting by column values on the block and address page tables was not working properly for some columns.
The issue was if there was an added suffix like "XEC" or "kB" for example it was breaking the plugin's sorting.
I modified some of the columns rendering functions to use a data attribute instead for displaying the suffixes

Test Plan: run the app and see if sorting works better

Reviewers: #bitcoin_abc, Mengerian

Reviewed By: #bitcoin_abc, Mengerian

Subscribers: Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D13083