[alias-server] [p9] Refactor functions to accept constants as input params
Summary:
T2922

Depends on D13084

Since there are no secrets in `config.js`, start including it in the repo. Refactor functions in `alias.js` to accept alias constants as input params. This will allow unit tests to be created and preserved for test registration addresses and fee structures (and the production registration address and fee structure).

Test Plan: `npm test` and all pass

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D13111