Introduce facility to evaluate miner fund as a block parking policy
Summary:
This patch introduces MinerFundPolicy as a type of ParkingPolicy. These
policies can be called for a given block to determine if they pass various
checks. If not, the block can be parked and Avalanche voting will determine
its fate.

Unit tests are provided. Integration into validation code will come in a later
patch.

Test Plan:
  ninja check

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13129