Dramatically reduce mempool_updatefromblock.py duration
Summary:
The test intent is to check all the blocks txs are added back to the mempool in case of a reorg. It also checks for a weird edge case behavior that bypasses the number of descendants mempool limit during reorg.
But this test doesn't need to use so many transactions: we can take advantage of the -limitdescendantcount option to speed up the test, which is exactly what this diff does. Note that the ancestor limit is kept greater than the default so we keep checking that the option is properly enforced.

This takes down the time from 128s to 12s on my machine and moves the test out of the extended suite.

Depends on D13131.

Test Plan:
  ./test/functional/test_runner.py mempool_updatefromblock

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13132